rasa-bot/requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==25.1.0
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.9.1
aiormq==6.4.2
aiosignal==1.4.0
anyio==4.11.0
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
babel==2.17.0
bidict==0.23.1
botocore==1.40.69
CacheControl==0.12.14
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.12.2
cryptography==46.0.3
cycler==0.12.1
dask==2022.10.2
dnspython==2.3.0
docopt==0.6.2
exceptiongroup==1.3.0
fbmessenger==6.0.0
filelock==3.20.0
fire==0.7.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
jax==0.4.30
jaxlib==0.4.30
jmespath==1.0.1
joblib==1.5.2
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.9
libclang==18.1.1
locket==1.0.0
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.5.3
mattermostwrapper==2.2
ml_dtypes==0.5.3
msgpack==1.1.2
multidict==5.2.0
networkx==2.6.3
numpy==1.23.5
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==20.9
pamqp==3.2.1
partd==1.4.2
pillow==12.0.0
pluggy==1.6.0
portalocker==2.10.1
prompt-toolkit==3.0.28
propcache==0.4.1
protobuf==4.23.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==1.10.9
pydot==1.4.2
PyJWT==2.10.1
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.2.5
pyreadline3==3.5.4
pyrsistent==0.20.0
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-engineio==4.12.3
pytz==2022.7.1
pywin32==311
PyYAML==6.0.3
questionary==1.10.0
randomname==0.1.5
rasa-sdk==3.6.2
redis==4.6.0
regex==2022.10.31
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rocketchat-API==1.30.0
rsa==4.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
sentry-sdk==1.14.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sklearn-crfsuite==0.3.6
slack_sdk==3.37.0
sniffio==1.3.1
SQLAlchemy==1.4.54
structlog==23.3.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.4
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
terminaltables==3.1.10
threadpoolctl==3.6.0
toolz==1.1.0
tqdm==4.67.1
typer-slim==0.20.0
typing-utils==0.1.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.1.3
wrapt==1.14.2
wsproto==1.2.0
yarl==1.22.0
rasa-sdk>=3.6,<4
google-generativeai>=0.7
python-dotenv>=1.0
requests>=2.0